index.html, styles.css, scripts.js, README.md
index.html
Added the div for the modal container.

styles.css
Added styling to the modal including a simple zoom animation. Added responsiveness to the modal.

scripts.js
Added code for the modal to open and close via clicks on screen, a close button and the escape key on the keyboard. Also displayed the height and image of the pokemon in the modal.

README.md
Updated some information for the use of the app and to view it remotely